fix: split edit tests for windows support
