+++ pkg/editor/edit_test.go
@@ -3,18 +3,13 @@ package editor
 import (
 	"bytes"
 	"context"
-	"flag"
 	"os"
-	"os/exec"
 	"testing"
 
 	"github.com/gopasspw/gopass/pkg/ctxutil"
 	"github.com/gopasspw/gopass/pkg/out"
-	"github.com/gopasspw/gopass/tests/gptest"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	"github.com/urfave/cli"
 )
 
 func TestEdit(t *testing.T) {
@@ -31,55 +26,4 @@ func TestEdit(t *testing.T) {
 	_, err := Invoke(ctx, "true", []byte{})
 	assert.Error(t, err)
 	buf.Reset()
-}
-
-func TestEditor(t *testing.T) {
-	u := gptest.NewUnitTester(t)
-	defer u.Remove()
-
-	ctx := context.Background()
-	touch, err := exec.LookPath("touch")
-	require.NoError(t, err)
-
-	want := "foobar"
-	out, err := Invoke(ctx, touch, []byte(want))
-	require.NoError(t, err)
-	if string(out) != want {
-		t.Errorf("'%s' != '%s'", string(out), want)
-	}
-}
-
-func TestGetEditor(t *testing.T) {
-	app := cli.NewApp()
-
-	// --editor=fooed
-	fs := flag.NewFlagSet("default", flag.ContinueOnError)
-	sf := cli.StringFlag{
-		Name:  "editor",
-		Usage: "editor",
-	}
-	require.NoError(t, sf.ApplyWithError(fs))
-	require.NoError(t, fs.Parse([]string{"--editor", "fooed"}))
-	c := cli.NewContext(app, fs, nil)
-
-	assert.Equal(t, "fooed", Path(c))
-
-	// EDITOR
-	fs = flag.NewFlagSet("default", flag.ContinueOnError)
-	c = cli.NewContext(app, fs, nil)
-	assert.NoError(t, os.Setenv("EDITOR", "fooenv"))
-	assert.Equal(t, "fooenv", Path(c))
-	assert.NoError(t, os.Unsetenv("EDITOR"))
-
-	// editor
-	pathed, err := exec.LookPath("editor")
-	if err == nil {
-		assert.Equal(t, pathed, Path(c))
-	}
-
-	// vi
-	op := os.Getenv("PATH")
-	assert.NoError(t, os.Setenv("PATH", "/tmp"))
-	assert.Equal(t, "vi", Path(c))
-	assert.NoError(t, os.Setenv("PATH", op))
 }
